Timer Script and Statue Collision
1. Have added the Canvas Prefab but need to fix its size and position.
Right on left click the timer starts and you can see in console the time
decrementing. Also have given an if else option to restart when time
hits '0' and when the time goes less than 30 seconds the color goes red

2. Created 'SoldierCollider' tag when the laser hits it the soldier
flares up. Create different tags for different statues and if else them
as in the laser script

3. Added the timer and laser script in another folder at DON'T
PANIC\Assets\Scripts\LaserAndTimerScript as I thought in the updates u
have made u might have made some changes to these scripts. Delete the
folder and replace the previous files if you havent.